+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:50123/","mongoDBName":"jest"}
+{"mongoUri":"mongodb://127.0.0.1:64716/","mongoDBName":"jest"}
