+++ src/main/factories/login/login-validation.spec.ts
@@ -1,8 +1,6 @@
-import { EmailValidation } from '@presentation/helpers/validators/email-validation'
-import { RequiredFieldValidation } from '@presentation/helpers/validators/required-field-validation'
-import { Validation } from '@presentation/protocols/validation'
-import { ValidationComposite } from '@presentation/helpers/validators/validation-composite'
+import { RequiredFieldValidation, EmailValidation, ValidationComposite } from '@presentation/helpers/validators'
 import { EmailValidator } from '@presentation/protocols/email-validator.interface'
+import { Validation } from '@presentation/protocols/validation'
 import { makeLoginValidation } from './login-validation'
 
 jest.mock('@presentation/helpers/validators/validation-composite')
