+++ src/main/factories/login/login-validation.ts
@@ -1,7 +1,5 @@
-import { EmailValidation } from '@presentation/helpers/validators/email-validation'
-import { RequiredFieldValidation } from '@presentation/helpers/validators/required-field-validation'
+import { EmailValidation, RequiredFieldValidation, ValidationComposite } from '@presentation/helpers/validators'
 import { Validation } from '@presentation/protocols/validation'
-import { ValidationComposite } from '@presentation/helpers/validators/validation-composite'
 import { EmailValidatorAdapter } from '@presentation/utils/email-validator-adapter'
 
 export const makeLoginValidation = (): ValidationComposite => {
