+++ src/main/factories/signup/signup-validation.spec.ts
@@ -1,9 +1,11 @@
-import { CompareFieldsValidation } from '@presentation/helpers/validators/compare-fields-validation'
-import { EmailValidation } from '@presentation/helpers/validators/email-validation'
-import { RequiredFieldValidation } from '@presentation/helpers/validators/required-field-validation'
-import { Validation } from '@presentation/protocols/validation'
-import { ValidationComposite } from '@presentation/helpers/validators/validation-composite'
+import {
+  RequiredFieldValidation,
+  CompareFieldsValidation,
+  EmailValidation,
+  ValidationComposite
+} from '@presentation/helpers/validators'
 import { EmailValidator } from '@presentation/protocols/email-validator.interface'
+import { Validation } from '@presentation/protocols/validation'
 import { makeSignUpValidation } from './signup-validation'
 
 jest.mock('@presentation/helpers/validators/validation-composite')
